+++ src/jtree.ts
@@ -1,7 +1,14 @@
 import TreeUtils from "./base/TreeUtils"
 import TreeNode from "./base/TreeNode"
 
-import { GrammarProgram, GrammarBackedRootNode, GrammarBackedTerminalNode, GrammarBackedNonTerminalNode } from "./GrammarLanguage"
+import {
+  GrammarProgram,
+  GrammarBackedRootNode,
+  GrammarBackedTerminalNode,
+  GrammarBackedNonTerminalNode,
+  GrammarBackedBlobNode,
+  GrammarBackedErrorNode
+} from "./GrammarLanguage"
 import UnknownGrammarProgram from "./tools/UnknownGrammarProgram"
 import TreeNotationCodeMirrorMode from "./codemirror/TreeNotationCodeMirrorMode"
 
@@ -10,6 +17,8 @@ class jtree {
   static Utils = TreeUtils
   static TreeNode = TreeNode
   static NonTerminalNode = GrammarBackedNonTerminalNode
+  static BlobNode = GrammarBackedBlobNode
+  static ErrorNode = GrammarBackedErrorNode
   static TerminalNode = GrammarBackedTerminalNode
   static GrammarProgram = GrammarProgram
   static UnknownGrammarProgram = UnknownGrammarProgram